requirements.txt
pytz
django
psycopg2-binary
djangorestframework
django-cors-headers
django-apscheduler
python-dateutil
requests